requirements.txt
streamlit==1.31.0
pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pymupdf>=1.22.0
paddlepaddle
google-generativeai
